refactor(macos): split Ghostty platform adapters
